switch around experiment calculation methods
