+++ ee/clickhouse/queries/experiments/trend_experiment_result.py
@@ -3,10 +3,10 @@ from datetime import datetime
 from math import exp, lgamma, log
 from typing import List, Optional, Tuple, Type
 
+from numpy.random import default_rng
 from rest_framework.exceptions import ValidationError
 
 from ee.clickhouse.queries.trends.clickhouse_trends import ClickhouseTrends
-from ee.clickhouse.queries.util import logbeta
 from posthog.models.feature_flag import FeatureFlag
 from posthog.models.filters.filter import Filter
 from posthog.models.team import Team
@@ -19,6 +19,7 @@ CONTROL_VARIANT_KEY = "control"
 class Variant:
     key: str
     count: int
+    exposure: int
 
 
 class ClickhouseTrendExperimentResult:
@@ -78,13 +79,15 @@ class ClickhouseTrendExperimentResult:
         # this assumes the Trend insight is Cumulative
         control_variant = None
         test_variants = []
+
         for result in insight_results:
             count = result["count"]
             breakdown_value = result["breakdown_value"]
             if breakdown_value == CONTROL_VARIANT_KEY:
-                control_variant = Variant(key=breakdown_value, count=int(count))
+                # by default, all variants exposed for same duration, so same exposure value
+                control_variant = Variant(key=breakdown_value, count=int(count), exposure=1)
             else:
-                test_variants.append(Variant(breakdown_value, int(count)))
+                test_variants.append(Variant(breakdown_value, int(count), 1))
 
         return control_variant, test_variants
 
@@ -104,74 +107,60 @@ class ClickhouseTrendExperimentResult:
         if len(test_variants) < 1:
             raise ValidationError("Can't calculate A/B test results for less than 2 variants", code="no_data")
 
-        # second calculation:
-        # https://www.evanmiller.org/bayesian-ab-testing.html#count_ab
-
-        control_count = 1 + control_variant.count
-        control_exposure = 1
-        test_exposure = 1  # by default, all variants exposed for same duration
-
-        test_counts_and_exposure = [(1 + variant.count, test_exposure) for variant in test_variants]
-
-        return calculate_probability_of_winning_for_each([(control_count, control_exposure), *test_counts_and_exposure])
+        return calculate_probability_of_winning_for_each([control_variant, *test_variants])
 
 
-def calculate_probability_of_winning_for_each(variants: List[Tuple[int, int]]) -> List[Probability]:
+def simulate_winning_variant_for_arrival_rates(target_variant: Variant, variants: List[Variant]) -> float:
+    random_sampler = default_rng()
+    simulations_count = 1_000_000
+
+    variant_samples = []
+    for variant in variants:
+        # Get `N=simulations` samples from a Beta distribution with alpha = prior_success + variant_sucess,
+        # and beta = prior_failure + variant_failure
+        samples = random_sampler.gamma(variant.count + 1, variant.exposure, simulations_count)
+        variant_samples.append(samples)
+
+    target_variant_samples = random_sampler.gamma(target_variant.count + 1, target_variant.exposure, simulations_count)
+
+    winnings = 0
+    variant_conversions = list(zip(*variant_samples))
+    for i in range(simulations_count):
+        if target_variant_samples[i] > max(variant_conversions[i]):
+            winnings += 1
+
+    return winnings / simulations_count
+
+
+def calculate_probability_of_winning_for_each(variants: List[Variant]) -> List[Probability]:
     """
     Calculates the probability of winning for each variant.
     """
     if len(variants) == 2:
         # simple case
-        probability = probability_B_beats_A_count_data(variants[0][0], variants[0][1], variants[1][0], variants[1][1])
+        probability = simulate_winning_variant_for_arrival_rates(variants[1], [variants[0]])
         return [1 - probability, probability]
 
     elif len(variants) == 3:
-        probability_third_wins = probability_C_beats_A_and_B_count_data(
-            variants[0][0], variants[0][1], variants[1][0], variants[1][1], variants[2][0], variants[2][1]
-        )
-        probability_second_wins = probability_C_beats_A_and_B_count_data(
-            variants[0][0], variants[0][1], variants[2][0], variants[2][1], variants[1][0], variants[1][1]
-        )
+        probability_third_wins = simulate_winning_variant_for_arrival_rates(variants[2], [variants[0], variants[1]])
+        probability_second_wins = simulate_winning_variant_for_arrival_rates(variants[1], [variants[0], variants[2]])
         return [1 - probability_third_wins - probability_second_wins, probability_second_wins, probability_third_wins]
 
+    elif len(variants) == 4:
+        probability_fourth_wins = simulate_winning_variant_for_arrival_rates(
+            variants[3], [variants[0], variants[1], variants[2]]
+        )
+        probability_third_wins = simulate_winning_variant_for_arrival_rates(
+            variants[2], [variants[0], variants[1], variants[3]]
+        )
+        probability_second_wins = simulate_winning_variant_for_arrival_rates(
+            variants[1], [variants[0], variants[2], variants[3]]
+        )
+        return [
+            1 - probability_fourth_wins - probability_third_wins - probability_second_wins,
+            probability_second_wins,
+            probability_third_wins,
+            probability_fourth_wins,
+        ]
     else:
         raise ValidationError("Can't calculate A/B test results for more than 4 variants", code="too_much_data")
-
-
-def probability_B_beats_A_count_data(A_count: int, A_exposure: int, B_count: int, B_exposure: int) -> Probability:
-    total: Probability = 0
-    for i in range(B_count):
-        total += exp(
-            i * log(B_exposure)
-            + A_count * log(A_exposure)
-            - (i + A_count) * log(B_exposure + A_exposure)
-            - log(i + A_count)
-            - logbeta(i + 1, A_count)
-        )
-
-    return total
-
-
-def probability_C_beats_A_and_B_count_data(
-    A_count: int, A_exposure: int, B_count: int, B_exposure: int, C_count: int, C_exposure: int
-) -> Probability:
-    total: Probability = 0
-
-    for k in range(B_count):
-        for l in range(A_count):
-            total += exp(
-                k * log(B_exposure)
-                + l * log(A_exposure)
-                + C_count * log(C_exposure)
-                - (k + l + C_count) * log(B_exposure + A_exposure + C_exposure)
-                + lgamma(k + l + C_count)
-                - lgamma(k + 1)
-                - lgamma(l + 1)
-                - lgamma(C_count)
-            )
-    return (
-        1
-        - probability_B_beats_A_count_data(C_count, C_exposure, A_count, A_exposure)
-        - probability_B_beats_A_count_data(C_count, C_exposure, B_count, B_exposure)
-        + total
-    )
