+++ ee/clickhouse/views/test/test_clickhouse_experiments.py
@@ -360,7 +360,7 @@ class ClickhouseTestFunnelExperimentResults(ClickhouseTestMixin, LicensedTestMix
 
         # Variant with test: Beta(2, 3) and control: Beta(3, 1) distribution
         # The variant has very low probability of being better.
-        self.assertAlmostEqual(response_data["probability"]["test"], 0.2619, places=3)
+        self.assertAlmostEqual(response_data["probability"]["test"], 0.2619, places=2)
 
     @snapshot_clickhouse_queries
     def test_experiment_flow_with_event_results_for_three_test_variants(self):
@@ -466,10 +466,10 @@ class ClickhouseTestFunnelExperimentResults(ClickhouseTestMixin, LicensedTestMix
         self.assertEqual(result[1][1]["count"], 1)
         self.assertEqual("test", result[1][1]["breakdown_value"][0])
 
-        self.assertAlmostEqual(response_data["probability"]["test"], 0.095, places=3)
-        self.assertAlmostEqual(response_data["probability"]["test_1"], 0.193, places=3)
-        self.assertAlmostEqual(response_data["probability"]["test_2"], 0.372, places=3)
-        self.assertAlmostEqual(response_data["probability"]["control"], 0.340, places=3)
+        self.assertAlmostEqual(response_data["probability"]["test"], 0.095, places=2)
+        self.assertAlmostEqual(response_data["probability"]["test_1"], 0.193, places=2)
+        self.assertAlmostEqual(response_data["probability"]["test_2"], 0.372, places=2)
+        self.assertAlmostEqual(response_data["probability"]["control"], 0.340, places=2)
 
 
 class ClickhouseTestTrendExperimentResults(ClickhouseTestMixin, LicensedTestMixin, APIBaseTest):
@@ -534,7 +534,7 @@ class ClickhouseTestTrendExperimentResults(ClickhouseTestMixin, LicensedTestMixi
 
         # Variant with test: Beta(2, 1) and control: Beta(3, 1) distribution
         # The variant has low probability of being better.
-        self.assertAlmostEqual(response_data["probability"]["test"], 0.313, places=3)
+        self.assertAlmostEqual(response_data["probability"]["test"], 0.313, places=2)
 
     @snapshot_clickhouse_queries
     def test_experiment_flow_with_event_results_for_three_test_variants(self):
@@ -617,6 +617,6 @@ class ClickhouseTestTrendExperimentResults(ClickhouseTestMixin, LicensedTestMixi
         self.assertEqual("test_2", result[2]["breakdown_value"])
 
         # test missing from results, since no events
-        self.assertAlmostEqual(response_data["probability"]["test_1"], 0.299, places=3)
-        self.assertAlmostEqual(response_data["probability"]["test_2"], 0.119, places=3)
-        self.assertAlmostEqual(response_data["probability"]["control"], 0.583, places=3)
+        self.assertAlmostEqual(response_data["probability"]["test_1"], 0.299, places=2)
+        self.assertAlmostEqual(response_data["probability"]["test_2"], 0.119, places=2)
+        self.assertAlmostEqual(response_data["probability"]["control"], 0.583, places=2)
